# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "ActorType": {
        "enum": [
          "GUEST",
          "OWNER",
          "ADMIN",
          "SYSTEM"
        ],
        "type": "string"
      },
...
      "CancelReservationDto": {
        "properties": {
          "reason": {
            "example": "Prośba gościa",
            "maxLength": 500,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateManualReservationDto": {
        "properties": {
          "checkIn": {
            "example": "2026-08-14",
            "format": "date",
            "type": "string"
          },
          "checkOut": {
            "description": "Dzień wyjazdu (`[)`)",
            "example": "2026-08-18",
            "format": "date",
            "type": "string"
          },
          "guest": {
            "$ref": "#/components/schemas/ReservationGuestInputDto"
          },
          "guestNotes": {
            "maxLength": 2000,
            "nullable": true,
            "type": "string"
          },
          "guestsCount": {
            "example": 2,
            "maximum": 99,
            "minimum": 1,
            "type": "number"
          },
          "ignoreMinNights": {
            "default": false,
            "description": "Pomija minimalny pobyt BR-03 (Q-01)",
            "type": "boolean"
          },
          "internalNotes": {
            "description": "Widoczna tylko w panelu",
            "maxLength": 2000,
            "nullable": true,
            "type": "string"
          },
          "roomId": {
            "description": "Pokój tego obiektu (inaczej 404)",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "roomId",
          "checkIn",
          "checkOut",
          "guestsCount",
          "guest"
        ],
        "type": "object"
      },
...
      "ErrorResponseDto": {
        "properties": {
          "code": {
            "description": "Stabilny kod maszynowy (UPPER_SNAKE_CASE). UI rozpoznaje błędy wyłącznie po nim.",
            "example": "RESERVATION_OVERLAP",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "Dane dodatkowe zależne od kodu. Dla VALIDATION_ERROR: `{ fields: [{ field, messages }] }`.",
            "example": {
              "conflictingReservationNumber": "KL-2026-000118"
            },
            "type": "object"
          },
          "error": {
            "description": "Nazwa kodu HTTP",
            "example": "Conflict",
            "type": "string"
          },
          "message": {
            "description": "Komunikat po polsku (pomocniczy)",
            "example": "Wybrany termin koliduje z inną rezerwacją.",
            "type": "string"
          },
          "path": {
            "description": "Ścieżka żądania",
            "example": "/api/v1/properties/2b1…/reservations",
            "type": "string"
          },
          "requestId": {
            "description": "Identyfikator żądania (także w nagłówku X-Request-Id)",
            "example": "6f0c2a7e-4c1b-4a8e-9a43-1d6f5f1c2b3a",
            "type": "string"
          },
          "statusCode": {
            "description": "Kod HTTP",
            "example": 409,
            "type": "number"
          },
          "timestamp": {
            "description": "Czas błędu (ISO 8601 UTC)",
            "example": "2026-08-01T10:15:30.000Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "code",
          "message",
          "path",
          "timestamp",
          "requestId"
        ],
        "type": "object"
      },
...
      "NightPriceRecordDto": {
        "properties": {
          "date": {
            "example": "2026-08-14",
            "format": "date",
            "type": "string"
          },
          "price": {
            "description": "Grosze",
            "example": 37000,
            "type": "number"
          }
        },
        "required": [
          "date",
          "price"
        ],
        "type": "object"
      },
...
      "PaginationMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "type": "number"
          },
          "totalItems": {
            "example": 134,
            "type": "number"
          },
          "totalPages": {
            "example": 7,
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ],
        "type": "object"
      },
...
      "ReservationDto": {
        "properties": {
          "cancellationReason": {
            "nullable": true,
            "type": "string"
          },
          "cancelledAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "cancelledBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ActorType"
              }
            ],
            "nullable": true
          },
          "checkIn": {
            "example": "2026-08-14",
            "format": "date",
            "type": "string"
          },
          "checkOut": {
            "description": "Dzień wyjazdu (zakres `[checkIn, checkOut)`)",
            "example": "2026-08-18",
            "format": "date",
            "type": "string"
          },
          "confirmedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "example": "PLN",
            "type": "string"
          },
          "events": {
            "description": "Historia (Q-05), od najstarszych",
            "items": {
              "$ref": "#/components/schemas/ReservationEventDto"
            },
            "type": "array"
          },
          "expiresAt": {
            "description": "Tylko dla PENDING (BR-07)",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "guest": {
            "$ref": "#/components/schemas/ReservationGuestDto"
          },
          "guestNotes": {
            "nullable": true,
            "type": "string"
          },
          "guestsCount": {
            "example": 3,
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "internalNotes": {
            "description": "Widoczna tylko w panelu",
            "nullable": true,
            "type": "string"
          },
          "nights": {
            "example": 4,
            "type": "number"
          },
          "number": {
            "example": "KL-2026-000123",
            "type": "string"
          },
          "priceBreakdown": {
            "description": "Zamrożone ceny nocy (BR-05)",
            "items": {
              "$ref": "#/components/schemas/NightPriceRecordDto"
            },
            "type": "array"
          },
          "propertyId": {
            "format": "uuid",
            "type": "string"
          },
          "room": {
            "$ref": "#/components/schemas/ReservationRoomRefDto"
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReservationSource"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReservationStatus"
              }
            ]
          },
          "totalPrice": {
            "description": "Grosze",
            "example": 164000,
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "description": "Do `PATCH` (BR-11)",
            "example": 1,
            "type": "number"
          }
        },
        "required": [
          "id",
          "number",
          "propertyId",
          "room",
          "checkIn",
          "checkOut",
          "nights",
          "guestsCount",
          "totalPrice",
          "currency",
          "status",
          "source",
          "expiresAt",
          "createdAt",
          "guest",
          "priceBreakdown",
          "guestNotes",
          "internalNotes",
          "confirmedAt",
          "cancelledAt",
          "cancelledBy",
          "cancellationReason",
          "version",
          "events",
          "updatedAt"
        ],
        "type": "object"
      },
      "ReservationEventDto": {
        "properties": {
          "actorName": {
            "description": "OWNER/ADMIN",
            "example": "Jan Nowak",
            "nullable": true,
            "type": "string"
          },
          "actorType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ActorType"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "description": "Np. `{ fields: [...] }` dla UPDATED, `{ reason }` dla CANCELLED",
            "nullable": true,
            "type": "object"
          },
          "type": {
            "enum": [
              "CREATED",
              "UPDATED",
              "CONFIRMED",
              "CANCELLED",
              "EXPIRED",
              "COMPLETED"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "actorType",
          "actorName",
          "createdAt",
          "payload"
        ],
        "type": "object"
      },
      "ReservationGuestDto": {
        "properties": {
          "email": {
            "example": "anna.kowalska@example.com",
            "nullable": true,
            "type": "string"
          },
          "firstName": {
            "example": "Anna",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "lastName": {
            "example": "Kowalska",
            "type": "string"
          },
          "phone": {
            "example": "+48 600 100 200",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone"
        ],
        "type": "object"
      },
      "ReservationGuestInputDto": {
        "properties": {
          "email": {
            "description": "Opcjonalny przy rezerwacji ręcznej (Q-03); istniejący gość z tym e-mailem jest aktualizowany (Q-04)",
            "example": "anna.kowalska@example.com",
            "nullable": true,
            "type": "string"
          },
          "firstName": {
            "description": "Wymagane bez `id`",
            "example": "Anna",
            "maxLength": 100,
            "type": "string"
          },
          "id": {
            "description": "Istniejący gość tego obiektu",
            "format": "uuid",
            "type": "string"
          },
          "lastName": {
            "description": "Wymagane bez `id`",
            "example": "Kowalska",
            "maxLength": 100,
            "type": "string"
          },
          "phone": {
            "example": "+48 600 100 200",
            "maxLength": 30,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ReservationListItemDto": {
        "properties": {
          "checkIn": {
            "example": "2026-08-14",
            "format": "date",
            "type": "string"
          },
          "checkOut": {
            "description": "Dzień wyjazdu (zakres `[checkIn, checkOut)`)",
            "example": "2026-08-18",
            "format": "date",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "example": "PLN",
            "type": "string"
          },
          "expiresAt": {
            "description": "Tylko dla PENDING (BR-07)",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "guest": {
            "$ref": "#/components/schemas/ReservationGuestRefDto"
          },
          "guestsCount": {
            "example": 3,
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nights": {
            "example": 4,
            "type": "number"
          },
          "number": {
            "example": "KL-2026-000123",
            "type": "string"
          },
          "propertyId": {
            "format": "uuid",
            "type": "string"
          },
          "room": {
            "$ref": "#/components/schemas/ReservationRoomRefDto"
          },
          "source": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReservationSource"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReservationStatus"
              }
            ]
          },
          "totalPrice": {
            "description": "Grosze",
            "example": 164000,
            "type": "number"
          }
        },
        "required": [
          "id",
          "number",
          "propertyId",
          "room",
          "guest",
          "checkIn",
          "checkOut",
          "nights",
          "guestsCount",
          "totalPrice",
          "currency",
          "status",
          "source",
          "expiresAt",
          "createdAt"
        ],
        "type": "object"
      },
      "ReservationPageDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ReservationListItemDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "ReservationRoomRefDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "example": "Domek Sosna",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
      "ReservationSource": {
        "enum": [
          "ONLINE",
          "MANUAL"
        ],
        "type": "string"
      },
      "ReservationStatus": {
        "enum": [
          "PENDING",
          "CONFIRMED",
          "CANCELLED",
          "EXPIRED",
          "COMPLETED"
        ],
        "type": "string"
      },
...
      "UpdateReservationDto": {
        "properties": {
          "checkIn": {
            "format": "date",
            "type": "string"
          },
          "checkOut": {
            "format": "date",
            "type": "string"
          },
          "guestNotes": {
            "maxLength": 2000,
            "nullable": true,
            "type": "string"
          },
          "guestsCount": {
            "maximum": 99,
            "minimum": 1,
            "type": "number"
          },
          "ignoreMinNights": {
            "default": false,
            "description": "Tylko rezerwacja MANUAL: pomija BR-03 przy zmianie terminu (Q-01)",
            "type": "boolean"
          },
          "internalNotes": {
            "description": "Można zmieniać w każdym statusie",
            "maxLength": 2000,
            "nullable": true,
            "type": "string"
          },
          "roomId": {
            "description": "Inny pokój tego obiektu",
            "format": "uuid",
            "type": "string"
          },
          "version": {
            "description": "Wersja z ostatniego odczytu (BR-11)",
            "example": 3,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "kl_refresh": {
        "in": "cookie",
        "name": "kl_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST API systemu rezerwacji Klucznik. Konwencje: docs/architecture/api-conventions.md",
    "title": "Klucznik API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/properties/{propertyId}/reservations": {
      "post": {
        "operationId": "Reservations_createManual",
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            },
            "description": "Nagłówek `Location: /api/v1/reservations/:id`"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "RESERVATION_OVERLAP (BR-01)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "CAPACITY_EXCEEDED, MIN_NIGHTS_NOT_MET, INVALID_STAY_DATES, ROOM_NOT_BOOKABLE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rezerwacja ręczna (od razu potwierdzona)",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/api/v1/reservations": {
      "get": {
        "operationId": "Reservations_list",
        "parameters": [
          {
            "description": "Numer strony (od 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Liczba elementów na stronie",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Sortowanie `pole:asc|desc`, wiele po przecinku. Dozwolone pola: checkIn, createdAt, number, totalPrice.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "checkIn:asc",
              "example": "checkIn:asc",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "propertyId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "roomId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Statusy po przecinku: PENDING, CONFIRMED, CANCELLED, EXPIRED, COMPLETED",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "example": "PENDING,CONFIRMED",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReservationSource"
            }
          },
          {
            "description": "Pobyt ma noc w `[from, to]`",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "≥ from",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Szukaj po nazwisku lub e-mailu gościa albo numerze (min. 2 znaki)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "example": "kowalska",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationPageDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista rezerwacji z filtrami, wyszukiwaniem i paginacją",
        "tags": [
          "reservations"
        ]
      }
    },
    "/api/v1/reservations/{id}": {
      "get": {
        "operationId": "Reservations_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Szczegóły rezerwacji z historią",
        "tags": [
          "reservations"
        ]
      },
      "patch": {
        "operationId": "Reservations_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "VERSION_CONFLICT (BR-11), RESERVATION_OVERLAP (BR-01), RESERVATION_NOT_EDITABLE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "CAPACITY_EXCEEDED, MIN_NIGHTS_NOT_MET, INVALID_STAY_DATES, ROOM_NOT_BOOKABLE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edycja rezerwacji (optimistic locking, Q-02)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/api/v1/reservations/{id}/cancel": {
      "post": {
        "operationId": "Reservations_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "INVALID_STATUS_TRANSITION (BR-06)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Anuluje rezerwację (dla PENDING: odrzuca)",
        "tags": [
          "reservations"
        ]
      }
    },
    "/api/v1/reservations/{id}/confirm": {
      "post": {
        "operationId": "Reservations_confirm",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": "INVALID_STATUS_TRANSITION (BR-06, także po wygaśnięciu: BR-07)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Potwierdza rezerwację oczekującą",
        "tags": [
          "reservations"
        ]
      }
    },
...
  },
  "servers": [],
...
}